Add --max-epochs-to-process arg to sync commands
Currently the database is only saved once, after the sync has fully
completed. However when multiple epochs are sunk there's a non-zero
chance of a random RPC server failure occurring, which will abort the
sync.

A better solution would be to incrementally save the database during the
sync but for now this flag serves as a workaround.