Use the old exchange credentials database if available.
The old exchange credentials database will be queried, but any new
credentials will be added to the keyring. Probably better not to mix
the two sets of credentials.  The assumption is that the user will add
existing credentials to keyring via the sys command line and remove
the old database, otherwise the old database will be used as long as
it exists.